Module sqlite3::access

Access to open sqlite3 database by filename.

The core module requires explicit authority to access files and such, following the principle of least authority.

This module provides the privileged functions to create such authorities.

TODO: move mod access to its own crate so that linking to sqlite3 doesn't bring in this ambient authority.

Modules

flags

Structs

ByFilename
Access to a database by filename

Functions

open
Open a database by filename.