Trait sqlite3::core::Access

pub trait Access {
    unsafe fn open(self, db: *mut *mut sqlite3) -> c_int;
}

Authorization to connect to database.

Required Methods

unsafe fn open(self, db: *mut *mut sqlite3) -> c_int

Open a database connection.

Whether or not an error occurs, allocate a handle and update db to point to it. return SQLITE_OK as c_int or set the errmsg of the db handle and return a relevant result code.