Struct sqlite3::cursor::Cursor

pub struct Cursor<'db> {
    // some fields omitted
}

The database cursor.

Unlike Database, Cursor is dependent of the active database connection and cannot be shared across tasks (non-Sendable). It is a compile time error that Cursor outlives Database either by Database escaping the block scope or by the task-shared reference to Database being expired.

Methods

impl<'db> Cursor<'db>

fn reset(&mut self) -> ResultCode

Resets a prepared SQL statement, but does not reset its bindings. See http://www.sqlite.org/c3ref/reset.html

fn clear_bindings(&mut self) -> ResultCode

Resets all bindings on a prepared SQL statement. See http://www.sqlite.org/c3ref/clear_bindings.html

fn step(&mut self) -> ResultCode

Evaluates a prepared SQL statement one ore more times. See http://www.sqlite.org/c3ref/step.html

fn step_row(&mut self) -> SqliteResult<Option<RowMap>>

fn get_blob<'a>(&'a mut self, i: isize) -> Option<&'a [u8]>

See http://www.sqlite.org/c3ref/column_blob.html

fn get_int(&mut self, i: isize) -> isize

See http://www.sqlite.org/c3ref/column_blob.html

fn get_i64(&mut self, i: isize) -> i64

See http://www.sqlite.org/c3ref/column_blob.html

fn get_f64(&mut self, i: isize) -> f64

See http://www.sqlite.org/c3ref/column_blob.html

fn get_text<'a>(&'a mut self, i: isize) -> Option<&'a str>

See http://www.sqlite.org/c3ref/column_blob.html

fn get_bind_index(&self, name: &str) -> isize

See http://www.sqlite.org/c3ref/bind_parameter_index.html

fn get_column_count(&self) -> isize

Returns the number of columns in a result set. See http://www.sqlite.org/c3ref/data_count.html

fn get_column_name<'a>(&'a self, i: isize) -> &'a str

Returns the name of the column with index i in the result set. See http://www.sqlite.org/c3ref/column_name.html

fn get_column_type(&self, i: isize) -> ColumnType

Returns the type of the column with index i in the result set. See http://www.sqlite.org/c3ref/column_blob.html

fn get_column_names(&self) -> Vec<String>

Returns the names of all columns in the result set.

fn bind_params(&mut self, values: &[BindArg]) -> ResultCode

fn bind_param(&mut self, i: isize, value: &BindArg) -> ResultCode

See http://www.sqlite.org/c3ref/bind_blob.html