Struct sqlite3::database::Database

pub struct Database {
    // some fields omitted
}

The database connection.

SQLite database is Sendable but not Copyable nor Syncable. Consequently, it can be shared through std::sync::Mutex across tasks (as it grants an exclusive access to the connection) but cannot be shared through std::sync::RWLock.

Methods

impl Database

fn get_errmsg(&self) -> String

Returns the error message of the the most recent call. See http://www.sqlite.org/c3ref/errcode.html

fn prepare<'db>(&'db self, sql: &str, _tail: &Option<&str>) -> SqliteResult<Cursor<'db>>

Prepares/compiles an SQL statement. See http://www.sqlite.org/c3ref/prepare.html

fn exec(&mut self, sql: &str) -> SqliteResult<bool>

Executes an SQL statement. See http://www.sqlite.org/c3ref/exec.html

fn get_changes(&self) -> isize

Returns the number of modified/inserted/deleted rows by the most recent call. See http://www.sqlite.org/c3ref/changes.html

fn get_last_insert_rowid(&self) -> i64

Returns the ID of the last inserted row. See http://www.sqlite.org/c3ref/last_insert_rowid.html

fn set_busy_timeout(&mut self, ms: isize) -> ResultCode

Sets a busy timeout. See http://www.sqlite.org/c3ref/busy_timeout.html