Trait sqlite3::Query

pub trait Query<F, T> where F: FnMut(&mut ResultRow) -> SqliteResult<T> {
    fn query<'stmt>(&'stmt mut self, values: &[&ToSql], txform: F) -> SqliteResult<QueryResults<'stmt, T, F>>;
}

Mix in query() convenience function.

Required Methods

fn query<'stmt>(&'stmt mut self, values: &[&ToSql], txform: F) -> SqliteResult<QueryResults<'stmt, T, F>>

Iterate over query results after binding parameters.

Each of the values is bound to the statement (using to_sql) and the statement is executed.

Returns an iterator over rows transformed by txform, which computes a value for each row (or an error).