Struct sqlite3::core::PreparedStatement

pub struct PreparedStatement {
    // some fields omitted
}

A prepared statement.

Methods

impl PreparedStatement

fn execute(&mut self) -> ResultSet

Begin executing a statement.

impl PreparedStatement

A compiled prepared statement that may take parameters. Note: "The leftmost SQL parameter has an index of 1."1

fn ignore_detail(&mut self)

Opt out of copies of error message details.

fn bind_null(&mut self, i: ParamIx) -> SqliteResult<()>

Bind null to a statement parameter.

fn bind_int(&mut self, i: ParamIx, value: i32) -> SqliteResult<()>

Bind an int to a statement parameter.

fn bind_int64(&mut self, i: ParamIx, value: i64) -> SqliteResult<()>

Bind an int64 to a statement parameter.

fn bind_double(&mut self, i: ParamIx, value: f64) -> SqliteResult<()>

Bind a double to a statement parameter.

fn bind_text(&mut self, i: ParamIx, value: &str) -> SqliteResult<()>

Bind a (copy of a) str to a statement parameter.

TODO: support binding without copying strings, blobs

fn bind_blob(&mut self, i: ParamIx, value: &[u8]) -> SqliteResult<()>

Bind a (copy of a) byte sequence to a statement parameter.

TODO: support binding without copying strings, blobs

fn clear_bindings(&mut self)

Clear all parameter bindings.

fn bind_parameter_count(&mut self) -> ParamIx

Return the number of SQL parameters. If parameters of the ?NNN form are used, there may be gaps in the list.

unsafe fn expose(&mut self) -> *mut sqlite3_stmt

Expose the underlying sqlite3_stmt struct pointer for use with the ffi module.

fn changes(&self) -> u64

Return the number of database rows that were changed or inserted or deleted by this statement if it is the most recently run on its database connection.

cf sqlite3_changes.